refactor: move tryListen to hub.js and add port scanner tests
